Stage_6 and Stage_7
We now try to build up Ajax application. One of the features of Ajax is
changing the pages without sending requests to the server. The
communication of the front-end side and the back-end side is done by the
scripts behind. As the program does not need to generate any pages, we
develop api calls. To send back the result to the fron-end side, we use
the Output Class to wrap up all data. In the script of the front-end, we
usually use
$.post(URL, postData, success_back_func(object){}, 'json')
then we can do something inside the success_back_func(object){}, and we
will use object to carry out the result from the api.
To communicate with the back-end side (server) without re-load the
pages, we usually write
$('#the_area').submit(function (evt) {
evt.preventDefault();
in order to stop the page jump to the page (the api, which will not
bring any pages). If we have to go another places later, we can use
window.location.href='URL' to the URL we want.  We can even add
element.html to achieve dynamic webpage after calculation. Do not forget
to set 'json' as the format in the front-end and the back-end.